Added shared to the main repo
took shared out of gitignore and added it into
the top level directory.
made symlinks to the shared folder in each of
the applications that need it. maybe should have
made those hard links... may be done in a later
commit. doesn't really need to.

need to update the install scripts to include
the shared folder libraries.